Add files via upload
C program to input a matrix in the usual format, convert it to a sparse matrix, store it, and store the transpose of the sparse matrix, and display and store the transpose in the usual format. 

Additional (Included in Comments): Converting into usual formatted Transpose from the Sparse simultaneously, Sorting Sparse Transpose Matrix using Bubble Sort according to the row as well as the column numbers of the usually formatted matrix. 

The code also contains input validations and an extensive explanation of the code via comments.