Add files via upload
C Program to find the product of two complex-valued matrices. Each complex-valued matrix is represented by a pair of real-valued matrices.

A complex-valued matrix X is represented by a pair of matrices (A, B) where A and B contain real values. Write a program that computes the product of two complex-valued matrices (A, B) and (C, D), where (A, B) * (C, D) = (A + iB)(C + iD) = (AC - BD) + i(AD + BC).
Determine the number of additions and multiplications if all the matrices are of order/size n x n.

Code contains: Input validations and extensive explanation of the code via comments wherever necessary. 

Additional:  If all the matrices are of size n x n then the total number of additions/subtractions = 2(n^2)(n-1) and the total number of multiplications = n^3.