Add files via upload
Prerequisite Knowledge: A priority queue is a queue in which the dequeue operation removes an item at the front of the queue, but the enqueue operation inserts items according to their priorities. A higher priority item is always enqueued before a lower priority item. An element that has the same priority as one or more elements in the queue is enqueued after all the elements with that priority.

Menu-driven implementation to perform the following operations on a priority queue (non-circular) of size n (implemented using array) depending on the user's choice (where the value of n is provided by the user as input).
1) Insert an element (provided by the user as input) to the queue according to its priority (also provided by the user as input). 
2) Delete an element from the queue. 
3) Display the contents of the queue with their corresponding priorities.

Code contains an extensive explanation of itself via comments and input validations too.